pyproject.toml
[tool.poetry]
name = "rephacheck"
version = "0.5.0"
description = "Health check for PostgreSQL cluster managed by repmgr"
authors = ["Nicolas KAROLAK <nicolas.karolak@ubicast.eu>"]
license = "LGPL-3.0+"
readme = "README.rst"
homepage = "https://github.com/UbiCastTeam/rephacheck"
repository = "https://github.com/UbiCastTeam/rephacheck"
classifiers=[
    "Development Status :: 4 - Beta",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: GNU Lesser General Public License v3 (LGPLv3)",
    "Natural Language :: English",
    "Operating System :: POSIX :: Linux",
    "Operating System :: Unix",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.5",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Topic :: Database",
    "Topic :: Internet :: Proxy Servers",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
]
keywords = "haproxy healthcheck postgresql repmgr"

[tool.poetry.dependencies]
python = "^3.5"
toml = "^0.10.0"
psycopg2-binary = "^2.7"
tornado = "^5.1"

[tool.poetry.dev-dependencies]
black = {version = "^18.3-alpha.0", python = "^3.6"}
pylint = "^2.1"
flake8 = "^3.6"
mypy = "^0.641.0"
ipython = "^7.0"

[tool.poetry.scripts]
rephacheck = "rephacheck.query:main"
rephaserver = "rephacheck.server:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
